move new_settings module to just settings
